backend/src/requirements.txt
fastapi==0.104.1
uvicorn==0.15.0
mysql-connector-python==8.2.0
pandas==1.5.0
seaborn==0.12.0
matplotlib==3.6.0
google-cloud-storage
google-auth